# contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: NextGen Market API
  description: |
    Enterprise-grade marketplace API platform for Iranian market
    
    ## ویژگی‌ها (Features)
    - Production-ready REST API
    - OpenAPI 3.1 specification
    - Semantic versioning (v1, v2, v3)
    - Full backwards compatibility
    - Persian/Jalali date support
    - Iranian Rial currency formatting
    
    ## احراز هویت (Authentication)
    این API از JWT RS256 برای احراز هویت استفاده می‌کند.
    توکن‌ها هر 15 دقیقه منقضی می‌شوند و باید با refresh token تمدید شوند.
    
    ## محدودیت نرخ (Rate Limiting)
    - کاربران عادی: 100 درخواست در دقیقه
    - کاربران احراز هویت شده: 1000 درخواست در دقیقه
    - مدیران: 10000 درخواست در دقیقه
  version: 3.1.0
  contact:
    name: پشتیبانی API
    email: api-support@nextgen.local
    url: https://docs.nextgen.local
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.nextgen.local/api/v3
    description: Production API (سرور اصلی)
    variables:
      region:
        default: ir
        enum:
          - ir
          - us
          - eu
  - url: https://api-staging.nextgen.local/api/v3
    description: Staging API (سرور تست)
  - url: http://localhost:3000/api/v3
    description: Local Development (توسعه محلی)
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List products (لیست محصولات)
      description: |
        دریافت لیست صفحه‌بندی شده محصولات با فیلتر
        Get paginated list of products with filtering
      operationId: listProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          example: 20
        - name: search
          in: query
          schema:
            type: string
          description: جستجو در نام و توضیحات محصول
          example: "شیرآلات"
        - name: category
          in: query
          schema:
            type: string
          example: "لوازم بهداشتی"
        - name: min_price
          in: query
          schema:
            type: number
            format: float
          description: حداقل قیمت (ریال)
          example: 1000000
        - name: max_price
          in: query
          schema:
            type: number
            format: float
          description: حداکثر قیمت (ریال)
          example: 50000000
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating, newest]
          example: "newest"
      responses:
        '200':
          description: List of products (لیست محصولات)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProducts'
              example:
                data:
                  - id: "660e8400-e29b-41d4-a716-446655440000"
                    name: "شیر روشویی مدل کلاسیک"
                    description: "شیر روشویی با طراحی کلاسیک و کیفیت عالی"
                    price: 12500000
                    price_formatted: "۱۲,۵۰۰,۰۰۰ ریال"
                    category: "شیرآلات"
                    sku: "SHR-CLS-001"
                    stock: 150
                    rating: 4.5
                    reviews_count: 89
                    created_at: "1402-08-01T09:00:00+03:30"
                  - id: "660e8400-e29b-41d4-a716-446655440001"
                    name: "توالت فرنگی مدل لوکس"
                    description: "توالت فرنگی با سیستم شستشوی پیشرفته"
                    price: 45000000
                    price_formatted: "۴۵,۰۰۰,۰۰۰ ریال"
                    category: "توالت فرنگی"
                    sku: "TLT-LUX-001"
                    stock: 45
                    rating: 4.8
                    reviews_count: 156
                    created_at: "1402-07-15T11:30:00+03:30"
                pagination:
                  page: 1
                  limit: 20
                  total: 1250
                  pages: 63
        '400':
          description: Invalid query parameters (پارامترهای جستجو نامعتبر)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    post:
      tags:
        - Products
      summary: Create product (ایجاد محصول)
      description: |
        ایجاد محصول جدید (فقط فروشنده)
        Create new product (seller only)
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            example:
              name: "شیر حمام مدل مدرن"
              description: "شیر حمام با طراحی مدرن و صرفه‌جویی در مصرف آب"
              price: 18500000
              category: "شیرآلات"
              sku: "SHR-MDR-002"
              stock: 100
              specifications:
                جنس: "استیل ضدزنگ"
                گارانتی: "۵ سال"
                کشور_سازنده: "ایران"
      responses:
        '201':
          description: Product created successfully (محصول با موفقیت ایجاد شد)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
              example:
                id: "660e8400-e29b-41d4-a716-446655440002"
                name: "شیر حمام مدل مدرن"
                description: "شیر حمام با طراحی مدرن و صرفه‌جویی در مصرف آب"
                price: 18500000
                price_formatted: "۱۸,۵۰۰,۰۰۰ ریال"
                category: "شیرآلات"
                sku: "SHR-MDR-002"
                stock: 100
                rating: 0
                reviews_count: 0
                created_at: "1402-09-25T14:00:00+03:30"
        '400':
          description: Invalid request data (داده‌های درخواست نامعتبر)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized (غیرمجاز)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - seller only (دسترسی ممنوع - فقط فروشنده)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "RBAC_002"
                message: "فقط فروشندگان می‌توانند محصول ایجاد کنند"
                messageEn: "Only sellers can create products"
...
components:
  schemas:
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: نام محصول
        description:
          type: string
          description: توضیحات محصول
        price:
          type: number
          format: float
          minimum: 0
          description: قیمت (ریال)
        price_formatted:
          type: string
          description: قیمت فرمت شده (ریال)
        price_toman:
          type: string
          description: قیمت به تومان
        category:
          type: string
          description: دسته‌بندی
        sku:
          type: string
          description: کد محصول
        stock:
          type: integer
          minimum: 0
          description: موجودی
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          description: امتیاز
        reviews_count:
          type: integer
          description: تعداد نظرات
        specifications:
          type: object
          additionalProperties:
            type: string
          description: مشخصات فنی
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              alt:
                type: string
        vendor:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required: [name, price, category, sku]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 200
          description: نام محصول
        description:
          type: string
          maxLength: 2000
          description: توضیحات محصول
        price:
          type: number
          format: float
          minimum: 0
          description: قیمت (ریال)
        category:
          type: string
          description: دسته‌بندی
        sku:
          type: string
          pattern: '^[A-Z0-9\-]{3,20}$'
          description: کد محصول
        stock:
          type: integer
          minimum: 0
          description: موجودی
        specifications:
          type: object
          additionalProperties:
            type: string
          description: مشخصات فنی
...
    PaginatedProducts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Order Schemas
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: شماره صفحه
        limit:
          type: integer
          description: تعداد در هر صفحه
        total:
          type: integer
          description: تعداد کل
        pages:
          type: integer
          description: تعداد صفحات

    # Error Schemas with Persian Descriptions
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: کد خطا
        message:
          type: string
          description: پیام خطا (فارسی)
        messageEn:
          type: string
          description: پیام خطا (انگلیسی)
        details:
          type: object
          description: جزئیات اضافی
        timestamp:
          type: string
          format: date-time
          description: زمان وقوع خطا
        path:
          type: string
          description: مسیر درخواست
        requestId:
          type: string
          description: شناسه درخواست
    ValidationError:
      type: object
      required: [code, message, fields]
      properties:
        code:
          type: string
          default: VAL_001
        message:
          type: string
          description: پیام خطا (فارسی)
        messageEn:
          type: string
          description: پیام خطا (انگلیسی)
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: خطاهای هر فیلد
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        توکن دسترسی JWT (اعتبار 15 دقیقه)
        JWT access token (15 min validity)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        کلید API برای احراز هویت سرویس به سرویس
        API key for service-to-service authentication

# Error Codes Reference (مرجع کدهای خطا)
...